+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:example/my_flutter_app_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:easy_localization/easy_localization.dart';
@@ -45,9 +46,17 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
   int counter = 0;
+  bool _gender = true;
+
   incrementCounter() {
     setState(() {
       counter++;
+    });
+  }
+
+  switchGender(bool val) {
+    setState(() {
+      _gender = val;
     });
   }
 
@@ -87,20 +96,45 @@ class _MyHomePageState extends State<MyHomePage> {
           ],
         ),
         body: Center(
-          child: new Column(
+          child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              new Text(AppLocalizations.of(context)
+              Spacer(
+                flex: 1,
+              ),
+              Text(
+                AppLocalizations.of(context)
+                    .gender('switch', _gender ? "female" : "male"),
+                style: TextStyle(
+                    color: Colors.grey.shade600,
+                    fontSize: 19,
+                    fontWeight: FontWeight.bold),
+              ),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  Icon(MyFlutterApp.male_1),
+                  Switch(value: _gender, onChanged: switchGender),
+                  Icon(MyFlutterApp.female_1),
+                ],
+              ),
+              Spacer(
+                flex: 1,
+              ),
+              Text(AppLocalizations.of(context)
                   .tr('msg', args: ['aissat', 'Flutter'])),
-              new Text(AppLocalizations.of(context).plural('clicked', counter)),
-              new FlatButton(
-                onPressed: () async {
+              Text(AppLocalizations.of(context).plural('clicked', counter)),
+              FlatButton(
+                onPressed: () {
                   incrementCounter();
                 },
-                child: new Text(AppLocalizations.of(context).tr('clickMe')),
+                child: Text(AppLocalizations.of(context).tr('clickMe')),
               ),
-              new Text(
+              Text(
                 AppLocalizations.of(context).tr('profile.reset_password.title'),
+              ),
+              Spacer(
+                flex: 2,
               ),
             ],
           ),